requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
apscheduler>=3.10.1
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.1.2
pydantic-settings>=2.0.0

